Add Evepraisal, margin and list choice parameters
Added Evepraisal market and price percent parameters to show the prices
of certain trade hub and at certain price percent.
Also added the sell-buy margin used by this application to distinguish
which item will go to the output list and two radio buttons to select
which one of the list (to sell order or buyback/insta-sell) you want.