Add enableApplyButton and saveSettings methods
`enableApplyButton` is straightforward, it will enable the JButton
`applyBtn` with `true` as an argument. It will be used by future
setting elements when their state have been changed.

`saveSettings` is not yet complete but it will at least disable the
JButton `applyBtn` when the function is called. Both `okButton` and
`applyButton` have it as part of their ActionListener.